fix: enforced maximum size for environment varaiable
