+++ options_test.go
@@ -4,7 +4,6 @@
 package instana
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -186,6 +185,7 @@ func TestInstanaTracingDisableEnvVar(t *testing.T) {
 		name     string
 		envValue string
 		expected map[string]bool
+		tooLarge bool
 	}{
 		{
 			name:     "No env var",
@@ -199,6 +199,12 @@ func TestInstanaTracingDisableEnvVar(t *testing.T) {
 				"logging": true,
 			},
 		},
+		{
+			name:     "Value exceeds size limit",
+			envValue: strings.Repeat("x", MaxEnvValueSize+1),
+			expected: map[string]bool{},
+			tooLarge: true,
+		},
 	}
 
 	for _, tt := range tests {
@@ -209,6 +215,13 @@ func TestInstanaTracingDisableEnvVar(t *testing.T) {
 			}
 
 			opts := DefaultOptions()
+
+			// For the too large case, we expect the environment variable to be ignored
+			// and no categories to be disabled
+			if tt.tooLarge {
+				assert.Empty(t, opts.Tracer.DisableSpans, "Expected no disabled spans for too large env value")
+				return
+			}
 
 			// Check if the maps have the expected values
 			for k, v := range tt.expected {
@@ -329,136 +342,5 @@ func verifyDisabledCategories(t *testing.T, disableMap map[string]bool, expected
 		if !disableMap[category] {
 			t.Errorf("Expected category %q to be disabled, but it wasn't", category)
 		}
-	}
-}
-
-func TestValidateFile(t *testing.T) {
-	tempDir := t.TempDir()
-
-	validFilePath := filepath.Join(tempDir, "valid.txt")
-	err := os.WriteFile(validFilePath, []byte("test content"), 0644)
-	if err != nil {
-		t.Fatalf("Failed to create test file: %v", err)
-	}
-
-	tests := []struct {
-		name          string
-		getPathFn     func() (string, error)
-		expectedError bool
-		errorContains string
-	}{
-		{
-			name: "Valid file",
-			getPathFn: func() (string, error) {
-				return validFilePath, nil
-			},
-			expectedError: false,
-		},
-		{
-			name: "Non-existent file",
-			getPathFn: func() (string, error) {
-				return filepath.Join(tempDir, "nonexistent.txt"), nil
-			},
-			expectedError: true,
-			errorContains: "no such file or directory",
-		},
-		{
-			name: "Symlink to valid file",
-			getPathFn: func() (string, error) {
-				symlinkPath := filepath.Join(tempDir, "symlink.txt")
-				err = os.Symlink(validFilePath, symlinkPath)
-				if err != nil {
-					return "", fmt.Errorf("Skipping symlink test, could not create symlink: %v", err)
-				}
-				return symlinkPath, nil
-			},
-			expectedError: false,
-		},
-		{
-			name: "Directory instead of file",
-			getPathFn: func() (string, error) {
-				dirPath := filepath.Join(tempDir, "testdir")
-				err = os.Mkdir(dirPath, 0755)
-				if err != nil {
-					return "", fmt.Errorf("Failed to create test directory: %v", err)
-				}
-				return dirPath, nil
-			},
-			expectedError: true,
-			errorContains: "not a regular file",
-		},
-		{
-			name: "File too large",
-			getPathFn: func() (string, error) {
-				fpath := filepath.Join(tempDir, "big.conf")
-				f, err := os.Create(fpath)
-				if err != nil {
-					return "", fmt.Errorf("Failed to create test file: %v", err)
-				}
-				defer f.Close()
-				err = f.Truncate(1024*1024 + 1) // >1MB
-				if err != nil {
-					return "", fmt.Errorf("Failed to truncate test file: %v", err)
-				}
-				return fpath, nil
-			},
-			expectedError: true,
-			errorContains: "config file too large",
-		},
-		{
-			name: "World-readable file",
-			getPathFn: func() (string, error) {
-				worldReadablePath := filepath.Join(tempDir, "world-readable.txt")
-				err = os.WriteFile(worldReadablePath, []byte("world-readable content"), 0644)
-				if err != nil {
-					return "", fmt.Errorf("Failed to create world-readable test file: %v", err)
-				}
-				return worldReadablePath, nil
-			},
-			expectedError: false, // This should not error, but will log a warning
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			path, err := tt.getPathFn()
-			if err != nil {
-				t.Skip(err)
-			}
-			absPath, err := validateFile(path)
-
-			if (err != nil) != tt.expectedError {
-				if tt.expectedError {
-					t.Errorf("Expected error but got none")
-				} else {
-					t.Errorf("Expected no error but got: %v", err)
-				}
-			}
-
-			// If am error is expected, check that it contains the expected text
-			if tt.expectedError && err != nil && tt.errorContains != "" {
-				if !strings.Contains(err.Error(), tt.errorContains) {
-					t.Errorf("Error message '%s' does not contain '%s'", err.Error(), tt.errorContains)
-				}
-			}
-
-			// For successful cases, check that the returned path is absolute
-			if !tt.expectedError && err == nil {
-				if !filepath.IsAbs(absPath) {
-					t.Errorf("Expected absolute path, got: %s", absPath)
-				}
-
-				// For the symlink case, verify it was resolved
-				if tt.name == "Symlink to valid file" {
-					// The resolved path should be an absolute path to the target file
-					// Note: On macOS, /var/folders may resolve to /private/var/folders
-					// so just check that the base filename matches
-					if filepath.Base(absPath) != filepath.Base(validFilePath) {
-						t.Errorf("Symlink not properly resolved. Got: %s, Expected file with basename: %s",
-							absPath, filepath.Base(validFilePath))
-					}
-				}
-			}
-		})
 	}
 }
